requirements.txt
minemeld-core>=0.10.0a1
requests
